requirements.txt
Django>=5.0,<6.0
gunicorn
whitenoise
python-dotenv
google-generativeai>=0.7.0
